// backend/apps/depot-service/src/depot/enums/depot.enums.ts
export enum VehicleStatus {
  AVAILABLE = 'AVAILABLE',
  CREW_ASSIGNED = 'CREW_ASSIGNED',
  PLATFORM_ASSIGNED = 'PLATFORM_ASSIGNED',
  READY_FOR_DISPATCH = 'READY_FOR_DISPATCH',
  DISPATCHED = 'DISPATCHED',
  RUNNING = 'RUNNING',
  ARRIVED = 'ARRIVED',
  PARKED = 'PARKED',
  UNDER_MAINTENANCE = 'UNDER_MAINTENANCE',
}

export enum PlatformStatus {
  EMPTY = 'EMPTY',
  OCCUPIED = 'OCCUPIED',
  RESERVED = 'RESERVED',
}